requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
torch>=1.9.0
torchvision>=0.10.0
python-multipart>=0.0.5
Pillow>=8.3.1
opencv-python>=4.5.3
numpy>=1.21.2
